monkey.rb і dolphin.rb розділено на бібліотеки та головний файл для виконання
monkey.rb розділено на primate.rb і chimp.rb, і dolphin.rb розділено на mammal.rb і orka.rb, з яких orka і chimp виконують програми.
