+++ ruby_classes/chimp.rb
@@ -1,0 +1,7 @@
+#!/usr/bin/ruby
+
+require_relative './lib/monkey'
+
+albert = Primate.new("monkey")
+
+puts(albert.family)
