+++ ruby_classes/lib/dolphin.rb
@@ -1,21 +1,7 @@
-#!/usr/bin/ruby
-
-
-class Mammal
- 	def initialize(genus)
-		@genus = genus
-	end
-  def genus
-    return(@genus)
-  end
-end
+require_relative 'mammal'
 
 class Dolphin < Mammal
   def initialize
     super "dolphin"
   end
 end
-
-winter = Dolphin.new
-
-puts(winter.genus)
