+++ ruby_classes/lib/mammal.rb
@@ -1,0 +1,8 @@
+class Mammal
+ 	def initialize(genus)
+		@genus = genus
+	end
+  def genus
+    return(@genus)
+  end
+end
