+++ ruby_classes/lib/primate.rb
@@ -1,0 +1,8 @@
+class Primate
+  def initialize(family)
+    @family = family
+  end
+  def family
+    return(@family)
+  end
+end
