+++ ruby_classes/orka.rb
@@ -1,0 +1,7 @@
+#!/usr/bin/ruby
+
+require_relative './lib/dolphin'
+
+winter = Dolphin.new
+
+puts(winter.genus)
